external modules: refactored some sql utils modules, removed legacy rw_gu12 transform module
