+++ external/etl/file/directory.py
@@ -1,22 +1,13 @@
 import os
-from pathlib import Path
 from external.native.connections import get_engine
 from external.etl.file.excel import extract, is_valid
 from external.etl.sql import load_clean
+from external.utils import sql, fs
 from external.utils.sql.common import dummy_db_action, apply_on_db
 from external.utils.sql.delete import truncate_table
 from external.utils.sql.grant import grant_preset_priveleges
-from external.utils.sql.introspection import get_loaded_src_ids
+from external.utils.sql.introspection import check_missing_sources
 from external.utils.var import color
-
-
-def get_filepaths(dirpath):
-    dirpath = Path(dirpath)
-    if not dirpath.exists():
-        print(f"The provided path: {dirpath} does not exist.")
-        return None
-    files_roots = [file for file in dirpath.rglob('*')]
-    return files_roots
 
 
 def print_etl_debug_msg(config):
@@ -45,14 +36,6 @@ def print_etl_debug_msg(config):
     print(debug_msg)
 
 
-def check_missing_sources(sources, conn_id, table):
-    engine = get_engine(conn_id)
-    loaded = get_loaded_src_ids(engine, table)
-    missing = [path for path in sources if path not in loaded]
-
-    return missing
-
-
 def dir_extract(config):
 
     src = config['extract']['src']
@@ -63,11 +46,11 @@ def dir_extract(config):
     full_table_name = f"{sink['schema']}.{sink['table']}"
     mode = config['load']['mode']
 
-    filepaths = get_filepaths(dir_uri)
+    filepaths = fs.get_filepaths(dir_uri)
     sources = [f'{filepath}' for filepath in filepaths if is_valid(f'{filepath}')]
 
     if mode == 'incremental':
-        files_to_extract = check_missing_sources(sources, conn_id, full_table_name)
+        files_to_extract = sql.introspection.check_missing_sources(sources, conn_id, full_table_name)
     elif mode == 'replace':
         files_to_extract = sources
     else:
@@ -86,7 +69,6 @@ def dir_extract(config):
 
 
 def dir_load(meta, config):
-    from external.utils.sql.common import sql_prog_exc
 
     sink = config['load']['sink']
     conn_id = sink['conn_id']
@@ -98,8 +80,8 @@ def dir_load(meta, config):
     engine = get_engine(conn_id)
 
     db_actions_map = {
-        'incremental': [dummy_db_action],
-        'replace': [truncate_table]
+        'incremental': [sql.common.dummy_db_action],
+        'replace': [sql.delete.truncate_table]
     }
 
     try:
@@ -107,16 +89,16 @@ def dir_load(meta, config):
     except KeyError:
         db_on_start = []
         print(f'{color("WARNING:", "red")} Unknown load mode. Operation will be canceled.')
-    db_on_end = [grant_preset_priveleges]
+    db_on_end = [sql.grant.grant_preset_priveleges]
 
     try:
-        apply_on_db(engine=engine, table=full_table_name, actions=db_on_start)
-    except sql_prog_exc:
+        sql.apply_on_db(engine=engine, table=full_table_name, actions=db_on_start)
+    except sql.common.sql_prog_exc:
         print(f'Table {color(table), "yellow"} does not exist and will be created by loader.')
 
     for batch_meta in meta:
         if batch_meta is not None:
             load_clean(meta=batch_meta, config=config)
 
-    apply_on_db(engine=engine, table=full_table_name, actions=db_on_end)
+    sql.apply_on_db(engine=engine, table=full_table_name, actions=db_on_end)
