--- external/etl/transform/meta.py
@@ -1,3 +1,0 @@
-
-
-
