+++ external/etl/transform/rw_gu12.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from external.etl.transform import schema, rows
+from external.etl.transform.pandas import schema, rows
 
 pd.set_option('mode.chained_assignment', None)
 
@@ -31,10 +31,10 @@ def remove_rows_with_missing_values(df, config):
 def transform(df, transform_config):
     column_name = transform_config['column_name']
     columns_list = transform_config['columns_list']
+    columns_to_check = transform_config['drop_na_rows']['columns_to_check']
     df_with_headers = schema.define_headers(df, search_col_name=column_name)
     df_with_cols_renamed = rename_columns(df_with_headers, transform_config)
     df_selected = df_with_cols_renamed[[column for column in df_with_cols_renamed.columns if column in columns_list]]
-    columns_to_check = transform_config['drop_na_rows']['columns_to_check']
     clean_df = rows.drop_na_rows(df_selected, columns_to_check=columns_to_check)
 
     return clean_df
