+++ external/utils/sql/delete.py
@@ -1,14 +1,14 @@
 from sqlalchemy import text
 from sqlalchemy.orm import sessionmaker
 
-from external.utils.sql.common import ping_table, _throw_warning, sql_execute, sql_prog_exc
+from external.utils.sql.common import ping_table, _throw_warning, execute, sql_prog_exc
 
 def truncate_table(engine, table, **kwargs):
     query = f'truncate table {table}'
     act = 'truncated'
     if ping_table(engine, table, **kwargs):
         _throw_warning(table, act=act)
-        sql_execute(engine, query)
+        execute(engine, query)
         print(f'Table {table} {act}.')
 
 
@@ -19,7 +19,7 @@ def drop_table_if_exists(engine, table, cascade=False, **kwargs):
         query = f'drop table if exists {table}'
     act = 'dropped'
     _throw_warning(table, act=act)
-    sql_execute(engine, query)
+    execute(engine, query)
     print(f'Table {table} {act}.')
 
 
@@ -33,7 +33,7 @@ def delete_records_by_period(engine, table, params):
         end = params['end']
         query = f"delete from {table} where {column} between '{start}' and '{end}';"
     try:
-        res = sql_execute(engine, query)
+        res = execute(engine, query)
     except sql_prog_exc:
         print(f'Table {table} does not exist and will be created by loader.')
