+++ external/utils/sql/get.py
@@ -1,6 +1,6 @@
 
 import os
-from external.utils.sql.common import sql_execute
+from external.utils.sql.common import execute
 from external.utils.configparse import read_query_file
 
 
@@ -12,7 +12,7 @@ def get_all(engine, table, limit=None, **kwargs):
             f'from {table} ' \
             + limit_exp \
             + ';'
-    result = sql_execute(engine, query)
+    result = execute(engine, query)
 
     return result
 
@@ -26,7 +26,7 @@ def get_by_period(engine, table, period_col, start, end, limit=None, **kwargs):
             f'where {period_col} between \'{start}\' and \'{end}\'' \
             + limit_exp \
             + ';'
-    result = sql_execute(engine, query)
+    result = execute(engine, query)
 
     return result
 
@@ -36,5 +36,5 @@ def get_by_custom_query(engine, query, **kwargs):
     if os.path.exists(query):
         query = read_query_file(query)
 
-    result = sql_execute(engine, query)
+    result = execute(engine, query)
     return result
