+++ external/utils/sql/introspection.py
@@ -1,6 +1,7 @@
 import sqlalchemy.exc
 
-from external.utils.sql.common import sql_execute
+from external.native.connections import get_engine
+from external.utils.sql.common import execute
 from external.utils.var import color
 
 
@@ -11,7 +12,7 @@ def get_columns(engine, schema, table, **kwargs):
     where TABLE_SCHEMA = \'{schema}\' 
       and TABLE_NAME = \'{table}\';
     """
-    result = sql_execute(engine, query=query)
+    result = execute(engine, query=query)
     return result
 
 
@@ -19,9 +20,17 @@ def get_loaded_src_ids(engine, table, **kwargs):
     query = f'select distinct record_source from {table};'
 
     try:
-        src_list = sql_execute(engine, query)
+        src_list = execute(engine, query)
         names = [name for name, in src_list]
         return names
     except sqlalchemy.exc.ProgrammingError:
         print(f'Table "{color(table, "yellow")}" does not exist! Filenames list is empty!')
         return []
+
+
+def check_missing_sources(sources, conn_id, table):
+    engine = get_engine(conn_id)
+    loaded = get_loaded_src_ids(engine, table)
+    missing = [path for path in sources if path not in loaded]
+
+    return missing
